add journey business test
